Produce a tidy dataframe of Scaling Up Nutrition Movement's progress reports from 2011 to 2017

Returns a tidy data frame of Scaling Up Nutrition Movement's published annual progress reports from 2011 to 2017 with five columns: text, which contains the text of the progress reports divided into elements of up to about 70 characters each, page, which contains the page number to which the text belongs to in the corresponding report, linenumber, which contains the linenumber where the text can be found in the corresponding report, chapter, which contains the chapter where the text can be found in the corresponding report, and year, which contains the year in which the corresponding report was released.

progress_reports()

Value

A data frame with 5 columns: text, page, linenumber, chapter, year.

Examples

progress_reports()

#> # A tibble: 36,807 x 5
#>    text                                            page linenumber chapter  year
#>    <chr>                                          <int>      <int>   <int> <int>
#>  1 Preface                                            3          1       0  2011
#>  2 One year ago, I joined a group of leaders in …     3          2       0  2011
#>  3 We set ourselves the ambitious target of subs…     3          3       0  2011
#>  4 1,000-day period of a child’s life, from preg…     3          4       0  2011
#>  5 The need for such an initiative is abundantly…     3          5       0  2011
#>  6 following prolonged drought in the Horn of Af…     3          6       0  2011
#>  7 support national authorities as they help vul…     3          7       0  2011
#>  8 security, and resist the impact of climatic a…     3          8       0  2011
#>  9 obesity, diabetes and heart disease in later …     3          9       0  2011
#> 10 on Non-Communicable Diseases especially timel…     3         10       0  2011
#> # … with 36,797 more rows