Scaling Up Nutrition (SUN) Movement Annual Progress Reports

Scaling Up Nutrition (SUN) http://scalingupnutrition.org is a global movement of 60 countries aiming to end malnutrition in all its forms. The SUN Movement. This package contains various datasets produced by the Movement and functions that aid in the manipulation and processing of these datasets. It also includes functions that support analysis and visualisation of these datasets. Finally, a Shiny application is included to allow user to interactively engage with the publicly-available SUN Movement datasets.

See also